refactor: renamed import of input mock file
